Extract common code to link-scripts module
